[FIX] Bad merge of responsive number 3
