+++ Configurateur/front/lib/screens/container-list/container_list.dart
@@ -12,8 +12,10 @@ import 'package:front/components/footer.dart';
 import 'package:front/components/items-information.dart';
 import 'package:front/network/informations.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:front/services/size_service.dart';
 import 'package:front/services/storage_service.dart';
 import 'package:front/services/theme_service.dart';
+import 'package:front/styles/globalStyle.dart';
 import 'package:front/styles/themes.dart';
 import 'package:http/http.dart' as http;
 import 'package:provider/provider.dart';
@@ -259,14 +261,22 @@ class _ContainerPageState extends State<ContainerPage> {
     return showDialog(
       context: context,
       builder: (BuildContext context) {
+        ScreenFormat screenFormat = SizeService().getScreenFormat(context);
         return StatefulBuilder(
           builder: (BuildContext context, StateSetter setState) {
             return AlertDialog(
-              title: const Text("Modifier un objet"),
+              title: Text(
+                "Modifier un objet",
+                style: TextStyle(
+                  fontSize: screenFormat == ScreenFormat.desktop
+                      ? desktopFontSize
+                      : tabletFontSize,
+                ),
+              ),
               content: Container(
-                height: 220.0,
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisSize: MainAxisSize.min,
                   children: [
                     const SizedBox(height: 10.0),
                     TextField(
@@ -277,7 +287,14 @@ class _ContainerPageState extends State<ContainerPage> {
                     const SizedBox(height: 10.0),
                     Row(
                       children: [
-                        const Text("Disponible"),
+                        Text(
+                          "Disponible",
+                          style: TextStyle(
+                            fontSize: screenFormat == ScreenFormat.desktop
+                                ? desktopFontSize
+                                : tabletFontSize,
+                          ),
+                        ),
                         Switch(
                           value: isAvailable,
                           onChanged: (bool newValue) {
@@ -303,6 +320,7 @@ class _ContainerPageState extends State<ContainerPage> {
                           labelText: "Nouvelle description",
                           hintText: initialDesc),
                     ),
+                    const SizedBox(height: 10.0),
                   ],
                 ),
               ),
@@ -311,7 +329,14 @@ class _ContainerPageState extends State<ContainerPage> {
                   onPressed: () {
                     Navigator.of(context).pop();
                   },
-                  child: const Text("Annuler"),
+                  child: Text(
+                    "Annuler",
+                    style: TextStyle(
+                      fontSize: screenFormat == ScreenFormat.desktop
+                          ? desktopFontSize
+                          : tabletFontSize,
+                    ),
+                  ),
                 ),
                 ElevatedButton(
                   onPressed: () async {
@@ -319,7 +344,14 @@ class _ContainerPageState extends State<ContainerPage> {
                         price, item, itemId);
                     Navigator.of(context).pop();
                   },
-                  child: const Text("Modifier"),
+                  child: Text(
+                    "Modifier",
+                    style: TextStyle(
+                      fontSize: screenFormat == ScreenFormat.desktop
+                          ? desktopFontSize
+                          : tabletFontSize,
+                    ),
+                  ),
                 ),
               ],
             );
@@ -412,6 +444,7 @@ class _ContainerPageState extends State<ContainerPage> {
     double price = 0.0;
     int selectedContainerId = 0;
     TextEditingController description = TextEditingController();
+    ScreenFormat screenFormat = SizeService().getScreenFormat(context);
 
     return showDialog(
       context: context,
@@ -419,11 +452,18 @@ class _ContainerPageState extends State<ContainerPage> {
         return StatefulBuilder(
           builder: (BuildContext context, StateSetter setState) {
             return AlertDialog(
-              title: const Text("Créer un nouvel objet"),
+              title: Text(
+                "Créer un nouvel objet",
+                style: TextStyle(
+                  fontSize: screenFormat == ScreenFormat.desktop
+                      ? desktopFontSize
+                      : tabletBigFontSize,
+                ),
+              ),
               content: Container(
-                height: 290.0,
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisSize: MainAxisSize.min,
                   children: [
                     const SizedBox(height: 10.0),
                     TextField(
@@ -434,7 +474,14 @@ class _ContainerPageState extends State<ContainerPage> {
                     const SizedBox(height: 10.0),
                     Row(
                       children: [
-                        const Text("Disponible"),
+                        Text(
+                          "Disponible",
+                          style: TextStyle(
+                            fontSize: screenFormat == ScreenFormat.desktop
+                                ? desktopFontSize
+                                : tabletBigFontSize,
+                          ),
+                        ),
                         Switch(
                           value: isAvailable,
                           onChanged: (bool newValue) {
@@ -476,6 +523,7 @@ class _ContainerPageState extends State<ContainerPage> {
                       decoration: const InputDecoration(
                           labelText: "Description de l'objet"),
                     ),
+                    const SizedBox(height: 10.0),
                   ],
                 ),
               ),
@@ -484,7 +532,14 @@ class _ContainerPageState extends State<ContainerPage> {
                   onPressed: () {
                     Navigator.of(context).pop();
                   },
-                  child: const Text("Annuler"),
+                  child: Text(
+                    "Annuler",
+                    style: TextStyle(
+                      fontSize: screenFormat == ScreenFormat.desktop
+                          ? desktopFontSize
+                          : tabletBigFontSize,
+                    ),
+                  ),
                 ),
                 ElevatedButton(
                   onPressed: () async {
@@ -493,7 +548,14 @@ class _ContainerPageState extends State<ContainerPage> {
 
                     Navigator.of(context).pop();
                   },
-                  child: const Text("Créer"),
+                  child: Text(
+                    "Créer",
+                    style: TextStyle(
+                      fontSize: screenFormat == ScreenFormat.desktop
+                          ? desktopFontSize
+                          : tabletBigFontSize,
+                    ),
+                  ),
                 ),
               ],
             );
@@ -504,6 +566,7 @@ class _ContainerPageState extends State<ContainerPage> {
   }
 
   Widget buildItemWidget(BuildContext context, ItemListInfo item) {
+    ScreenFormat screenFormat = SizeService().getScreenFormat(context);
     return GestureDetector(
       onTap: () {},
       child: Card(
@@ -512,10 +575,31 @@ class _ContainerPageState extends State<ContainerPage> {
           children: [
             Expanded(
               child: ListTile(
-                title: Text("nom : ${item.name}"),
+                title: Text(
+                  "nom : ${item.name}",
+                  style: TextStyle(
+                    fontSize: screenFormat == ScreenFormat.desktop
+                        ? desktopFontSize
+                        : tabletBigFontSize,
+                  ),
+                ),
                 subtitle: item.description != null
-                    ? Text("description : ${item.description!}")
-                    : Text("description : pas de description"),
+                    ? Text(
+                        "description : ${item.description!}",
+                        style: TextStyle(
+                          fontSize: screenFormat == ScreenFormat.desktop
+                              ? desktopFontSize
+                              : tabletBigFontSize,
+                        ),
+                      )
+                    : Text(
+                        "description : pas de description",
+                        style: TextStyle(
+                          fontSize: screenFormat == ScreenFormat.desktop
+                              ? desktopFontSize
+                              : tabletBigFontSize,
+                        ),
+                      ),
               ),
             ),
             Row(
@@ -550,6 +634,7 @@ class _ContainerPageState extends State<ContainerPage> {
   }
 
   Widget build(BuildContext context) {
+    ScreenFormat screenFormat = SizeService().getScreenFormat(context);
     return DefaultTabController(
       length: 2,
       child: Scaffold(
@@ -571,13 +656,22 @@ class _ContainerPageState extends State<ContainerPage> {
                     Tab(
                       child: Text(
                         'Liste des conteneurs',
-                        style: TextStyle(color: Colors.blue),
+                        style: TextStyle(
+                          color: Colors.blue,
+                          fontSize: screenFormat == ScreenFormat.desktop
+                              ? desktopFontSize
+                              : tabletBigFontSize,
+                        ),
                       ),
                     ),
                     Tab(
                       child: Text(
                         'Liste des objets',
-                        style: TextStyle(color: Colors.blue),
+                        style: TextStyle(
+                            color: Colors.blue,
+                            fontSize: screenFormat == ScreenFormat.desktop
+                                ? desktopFontSize
+                                : tabletBigFontSize),
                       ),
                     ),
                   ],
@@ -602,7 +696,9 @@ class _ContainerPageState extends State<ContainerPage> {
                             child: Text(
                               'Aucun conteneur trouvé.',
                               style: TextStyle(
-                                fontSize: 18,
+                                fontSize: screenFormat == ScreenFormat.desktop
+                                    ? desktopFontSize
+                                    : tabletBigFontSize,
                                 color: Color.fromARGB(255, 211, 11, 11),
                               ),
                             ),
@@ -635,7 +731,9 @@ class _ContainerPageState extends State<ContainerPage> {
                             child: Text(
                               'Aucun objet trouvé.',
                               style: TextStyle(
-                                fontSize: 18,
+                                fontSize: screenFormat == ScreenFormat.desktop
+                                    ? desktopFontSize
+                                    : tabletBigFontSize,
                                 color: Color.fromARGB(255, 211, 11, 11),
                               ),
                             ),
@@ -677,12 +775,20 @@ class _ContainerPageState extends State<ContainerPage> {
                                             BorderRadius.circular(20.0),
                                       ),
                                     ),
-                                    child: const Row(
+                                    child: Row(
                                       mainAxisSize: MainAxisSize.min,
                                       children: [
-                                        Icon(Icons.add),
-                                        SizedBox(width: 8),
-                                        Text('Ajouter un items'),
+                                        const Icon(Icons.add),
+                                        const SizedBox(width: 8),
+                                        Text(
+                                          'Ajouter un items',
+                                          style: TextStyle(
+                                            fontSize: screenFormat ==
+                                                    ScreenFormat.desktop
+                                                ? desktopFontSize
+                                                : tabletBigFontSize,
+                                          ),
+                                        ),
                                       ],
                                     ),
                                   ),
@@ -719,7 +825,10 @@ class _ContainerPageState extends State<ContainerPage> {
                                       child: Text(
                                         'Aucun objet trouvé.',
                                         style: TextStyle(
-                                          fontSize: 18,
+                                          fontSize: screenFormat ==
+                                                  ScreenFormat.desktop
+                                              ? desktopFontSize
+                                              : tabletBigFontSize,
                                           color:
                                               Color.fromARGB(255, 211, 11, 11),
                                         ),
