+++ Configurateur/front/lib/screens/feedbacks/feedbacks_style.dart
@@ -1,0 +1,7 @@
+import 'package:sizer/sizer.dart';
+
+var tabletContainerWidth = 40.0.w;
+var desktopContainerWidth = 30.0.w;
+
+var tabletContainerHeight = 50.0.h;
+var desktopContainerHeight = 60.0.h;
