+++ Configurateur/front/lib/screens/login/login_style.dart
@@ -1,0 +1,9 @@
+import 'package:sizer/sizer.dart';
+
+const tabletWidthFactor = 0.5;
+const desktopWidthFactor = 0.3;
+
+var desktopButtonWidth = 20.0.w;
+var desktopButtonHeight = 5.0.h;
+var tabletButtonWidth = 25.0.w;
+var tabletButtonHeight = 5.0.h;
