+++ Configurateur/front/lib/screens/password-recuperation/password_change_style.dart
@@ -1,0 +1,7 @@
+import 'package:sizer/sizer.dart';
+
+var tabletWidthFactor = 0.4;
+var desktopWidthFactor = 0.2;
+
+var tabletSendButtonWidth = 30.0.w;
+var desktopSendButtonWidth = 20.0.w;
