+++ Configurateur/front/lib/screens/password-recuperation/password_recuperation_style.dart
@@ -1,0 +1,12 @@
+import 'package:sizer/sizer.dart';
+
+var tabletWidthFactor = 0.4;
+var desktopWidthFactor = 0.2;
+
+var tabletCaptchaWidth = 30.0.w;
+var desktopCaptchaWidth = 20.0.w;
+var tabletCaptchaHeight = 10.0.h;
+var desktopCaptchaHeight = 7.0.h;
+
+var tabletSendButtonWidth = 30.0.w;
+var desktopSendButtonWidth = 20.0.w;
